fix(ts): resolve build errors after refactor
